fix(hmr): invalidate module with virtual module content, fixes #86
Previously the virtual module would be invalidated with some generic code which which could guaranteed to be unique.

For some reason in Next.js (webpack5) it would write the module with this source but not run it through the loader. While this doesn't fix the root cause it is a bandaid solution to avoid a broken state from the HMR.

Also we will no longer write the source details to the virtual module, it wasn't needed and was there for debugging purposes.

This also may fix duplicate HMR runs leading to high HMR times.