fix(webpack5): Fix null invalid file names causing HMR issues, fixes #86
For whatever reason in webpack 5, it seems that the invalid hook is called when you modify a file but it doesn't give you the file name. To get around this when we detect a null file change we re-scan all files. This will slow down HMR but there's not really a better alternative for now.

Can optimise the re-scanning at some point.